Move  asyncConnect to UsersControlpanel.jsx
