+++ packages/volto/src/components/manage/Controlpanels/Users/UsersControlpanel.test.jsx
@@ -38,6 +38,10 @@ describe('UsersControlpanel', () => {
         locale: 'en',
         messages: {},
       },
+      router: {
+        location: { pathname: '/controlpanel/users' },
+      },
+      reduxAsyncConnect: {},
     });
     const { container } = render(
       <Provider store={store}>
@@ -82,6 +86,10 @@ describe('UsersControlpanel', () => {
         locale: 'en',
         messages: {},
       },
+      router: {
+        location: { pathname: '/controlpanel/users' },
+      },
+      reduxAsyncConnect: {},
     });
 
     render(
