--- packages/volto/types/jwt-decode.d.ts
@@ -1,4 +1,0 @@
-declare module 'jwt-decode' {
-  export function jwtDecode<T = any>(token: string): T;
-  export default jwtDecode;
-}
